Fix deadlock caused by leaked loaded applications
A bug was present in the original PR (#2192). When running a query, the
`ServiceSyncRuntime::run_query` function didn't use the `SyncRuntime`
guard type, so the `Drop` implementation that cleared the applications
was never executed.

I missed this because on an upcoming PR the `run_query` associated
function becomes a method, so it forces the usage of the `SyncRuntime`
guard type, fixing the bug.